requirements.txt
Flask==2.3.2
werkzeug==2.3.2
inference-sdk==1.0.0
numpy==1.24.2
matplotlib==3.8.0
Pillow==10.0.1
requests==2.31.0
